Extract package_list out of install.rs.
We want to use this externally to walk the package tree so pull
out PackageInstall::package_list to core::package::list().
