+++ components/core/src/package/mod.rs
@@ -19,9 +19,12 @@ pub mod metadata;
 pub mod plan;
 pub mod target;
 
+mod list;
+
 pub use self::archive::{FromArchive, PackageArchive};
 pub use self::ident::{Identifiable, PackageIdent};
 pub use self::install::PackageInstall;
+pub use self::list::list;
 pub use self::plan::Plan;
 pub use self::target::PackageTarget;
